% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_shapley_interaction}
\alias{new_shapley_interaction}
\title{Class constructor for class \code{shapley_interaction}.}
\usage{
new_shapley_interaction(PHI = numeric())
}
\arguments{
\item{PHI}{A \eqn{p \times p} matrix containing the decomposition of the squared Mahalanobis distance
of a \eqn{p}-dimensional numeric vector into outlyingness scores for pairs of variables.}
}
\value{
Matrix of class \code{shapley_interaction}, containing input matrix \code{PHI}.
}
\description{
This function creates an object of class \code{shapley_interaction} that is returned
by the \code{\link{shapley_interaction}} function.
}
