% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shapley}
\alias{print.shapley}
\title{Print function for class \code{shapley}.}
\usage{
\method{print}{shapley}(x, ...)
}
\arguments{
\item{x}{List of class \code{shapley}.}

\item{...}{Optional arguments passed to methods.}
}
\value{
Prints the list entries of \code{x} that are not \code{NULL}.
}
\description{
Print function for class \code{shapley}.
}
