% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shapley_algorithm}
\alias{print.shapley_algorithm}
\title{Print function for class \code{shapley_algorithm}.}
\usage{
\method{print}{shapley_algorithm}(x, ...)
}
\arguments{
\item{x}{List of class \code{shapley_algorithm}.}

\item{...}{Optional arguments passed to methods.}
}
\value{
Prints the imputed data and the Shapley values.
}
\description{
Print function for class \code{shapley_algorithm}.
}
