% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shapley_interaction}
\alias{print.shapley_interaction}
\title{Print function for class \code{shapley_interaction}.}
\usage{
\method{print}{shapley_interaction}(x, ...)
}
\arguments{
\item{x}{Matrix of class \code{shapley_interaction}.}

\item{...}{Optional arguments passed to methods.}
}
\value{
Prints the Shapley interaction indices.
}
\description{
Print function for class \code{shapley_interaction}.
}
