% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapleyOutlier-package.R
\docType{package}
\name{ShapleyOutlier-package}
\alias{ShapleyOutlier}
\alias{ShapleyOutlier-package}
\title{ShapleyOutlier: Multivariate Outlier Explanations using Shapley Values and Mahalanobis Distances}
\description{
Based on Shapley values to explain multivariate outlyingness and to detect and impute cellwise outliers. Includes implementations of methods described in Mayrhofer and Filzmoser (2023) \doi{10.1016/j.ecosta.2023.04.003}.
}
\author{
\strong{Maintainer}: Marcus Mayrhofer \email{marcus.mayrhofer@tuwien.ac.at}

Authors:
\itemize{
  \item Peter Filzmoser
}

}
\keyword{internal}
